Update GraphQL HTTP methods and add mo library
The code has been refactored to include the mo library in the dependencies, leading to a restructuring of the HTTP methods in the GraphQL client. The error handling has been improved with the help of this library. A 'Query' function was added for formulating a request query and the 'Query' method which used to send a POST request was renamed to 'Do' method enhancing clarity.